+++ src/app/locales/en/utils/archive-filesystem-browser-model.js
@@ -1,0 +1,6 @@
+import FilesystemBrowserModel from './filesystem-browser-model';
+import _ from 'lodash';
+
+export default _.merge({}, FilesystemBrowserModel, {
+  checkingSymlinkInArchive: 'checking symbolic link',
+});