+++ src/app/utils/archive-filesystem-browser-model.js
@@ -23,6 +23,13 @@ import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 
 export default FilesystemBrowserModel.extend({
   modalManager: service(),
+  archiveManager: service(),
+  globalNotify: service(),
+
+  /**
+   * @override
+   */
+  i18nPrefix: 'utils.archiveFilesystemBrowserModel',
 
   /**
    * @virtual
@@ -268,15 +275,20 @@ export default FilesystemBrowserModel.extend({
    *   should be handled by question to user, not by standard dir change
    */
   async handlePotentialExternalSymlink(symlink) {
-    const externalContext = await this.symlinkExternalContext(symlink);
-    if (externalContext) {
-      this.openExternalSymlinkModal(
-        symlink,
-        externalContext,
-      );
+    try {
+      const externalContext = await this.symlinkExternalContext(symlink);
+      if (externalContext && !(await this.isNestedArchiveContext(externalContext))) {
+        this.openExternalSymlinkModal(
+          symlink,
+          externalContext,
+        );
+        return true;
+      } else {
+        return false;
+      }
+    } catch (error) {
+      this.get('globalNotify').backendError(this.t('checkingSymlinkInArchive'), error);
       return true;
-    } else {
-      return false;
     }
   },
 
@@ -322,5 +334,25 @@ export default FilesystemBrowserModel.extend({
       modalManager.hide(get(externalSymlinkModal, 'id'));
       this.set('externalSymlinkModal', null);
     }
+  },
+
+  async isNestedArchiveContext(filesViewContext) {
+    const {
+      archiveManager,
+      archive,
+    } = this.getProperties('archiveManager', 'archive');
+    const archiveId = get(filesViewContext, 'archiveId');
+    if (!archiveId) {
+      return false;
+    }
+    const archiveToOpen = await archiveManager.getArchive(archiveId);
+    const modelArchiveId = get(archive, 'entityId');
+    let checkedArchiveParent = archiveToOpen;
+    do {
+      checkedArchiveParent = await get(checkedArchiveParent, 'parentArchive');
+    } while (
+      checkedArchiveParent && get(checkedArchiveParent, 'entityId') !== modelArchiveId
+    );
+    return Boolean(checkedArchiveParent);
   },
 });